test(command-resolver): fold the findExecutableInEnv suite into commandResolver.test.ts
Test basenames must match the file under test (naming conventions
§3.3). The suite resets the module registry and re-imports with isWin
forced so the resolver-ordering regression still runs on every CI host,
which the host-platform skipIf suites in this file cannot.

Addresses review C2 (0xfullex) on #16402.
